agregar3

+++ src/main/java/com/example/demo/Calculadora.java
@@ -2,24 +2,24 @@ package com.example.demo;
 
 public class Calculadora {
 
-	private float num1;
-	private float num2;
+	private int num1;
+	private int num2;
 	private char operador;
 	private float resultado;
 	
-	public float getNum1() {
+	public int getNum1() {
 		return num1;
 	}
 	
-	public void setNum1(float num1) {
+	public void setNum1(int num1) {
 		this.num1 = num1;
 	}
 	
-	public float getNum2() {
+	public int getNum2() {
 		return num2;
 	}
 	
-	public void setNum2(float num2) {
+	public void setNum2(int num2) {
 		this.num2 = num2;
 	}
 	
@@ -52,7 +52,7 @@ public class Calculadora {
 			resultado=num1*num2;
 			break;
 		case '/':
-			resultado=num1/num2;
+			resultado=(float)num1 /(float)num2;
 			break;
 		default:
 			break;
